+++ src/view/pages/Statistics/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Statistics = () => {
+  return (
+    <>
+      <h1>Statistics</h1>
+    </>
+  )
+}
+
+export default Statistics;
