+++ src/view/pages/Team/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Team = () => {
+  return (
+    <>
+      <h1>Team</h1>
+    </>
+  )
+}
+
+export default Team;
